More or less finish bower2yarn scripting

+++ packages/bower2yarn/cli.js
@@ -1,5 +1,6 @@
 #!/usr/bin/env node
 
+const updateNotifier = require('update-notifier')
 const meow = require('meow')
 const bowerConfig = require('bower-config')
 const which = require('which')
@@ -8,6 +9,12 @@ const path = require('path')
 const fs = require('fs')
 const { startsWith } = require('lodash')
 const execa = require('execa')
+const cloneDeep = require('clone-deep')
+const difflet = require('difflet')({ indent: 2 })
+const deepIs = require('deep-is')
+
+const pkg = require('./package.json')
+updateNotifier({ pkg }).notify()
 
 const cli = meow(
   `
@@ -19,34 +26,25 @@ const cli = meow(
 )
 
 function help () {
-  console.log(cli.help)
+  console.error(cli.help)
   process.exit(1)
 }
 
-let i = 0
-const complete = []
+function step (title, lines, last = false) {
+  console.error()
+  console.error(chalk.green('# ' + title))
 
-function task (task) {
-  complete.push(`${++i}. ${task}`)
-}
-
-function step (lines, last = false) {
-  if (complete.length > 0) {
-    console.log()
-    console.log(chalk.green(complete.join('\n')))
-  }
-
-  console.log()
-  console.log(chalk.red(lines.join('\n')))
+  console.error()
+  console.error(lines.join('\n'))
 
   if (!last) {
-    console.log()
-    console.log(
-      chalk.gray("Please call bower2yarn once more when you're done with this!")
+    console.error()
+    console.error(
+      chalk.red("Please call bower2yarn once more when you're done with this!")
     )
   }
 
-  console.log()
+  console.error()
 
   process.exit(0)
 }
@@ -66,12 +64,26 @@ function exists (bin) {
 async function main () {
   const cwd = process.cwd()
 
-  task('Install Yarn')
+  if (!fs.existsSync(path.join(cwd, 'bower.json'))) {
+    if (fs.existsSync(path.join(cwd, 'node_modules', '@bower_components'))) {
+      step(
+        'Done',
+        ['Your project is now converted to Yarn! Thank you for using Bower!'],
+        true
+      )
+    } else {
+      step('Browse to project directory', [
+        'Current directory does not contain bower.json',
+        '',
+        'Please browse to directory that contains your project to convert.'
+      ])
+    }
+  }
 
   const yarnPath = exists('yarn', cwd)
 
   if (!yarnPath) {
-    step([
+    step('Install Yarn', [
       'A good first step to migrate to Yarn is installing it!',
       '',
       'Please choose your preferred method:',
@@ -87,12 +99,10 @@ async function main () {
     ])
   }
 
-  task('Install Bower with Yarn')
-
   const bowerPath = exists('bower')
 
   if (!bowerPath) {
-    step([
+    step('Install Bower', [
       'We cannot drop Bower just yet, we need it to install legacy dependencies.',
       '',
       'As a first step, please install Bower with:',
@@ -106,13 +116,14 @@ async function main () {
     ])
   }
 
-  task('Install dependencies')
-
   const config = bowerConfig.read(cwd)
   const componentsDir = path.resolve(cwd, config.directory)
 
-  if (!fs.existsSync(componentsDir)) {
-    step([
+  if (
+    !fs.existsSync(componentsDir) &&
+    !fs.existsSync(path.join(cwd, 'node_modules', '@bower_components'))
+  ) {
+    step('Install dependencies with Bower', [
       'We need to install dependencies the old way first. Please run:',
       '$ bower install',
       '',
@@ -121,9 +132,21 @@ async function main () {
     ])
   }
 
-  task('Create or update package.json')
+  let original = {}
 
-  const expected = {}
+  if (fs.existsSync('package.json')) {
+    original = JSON.parse(fs.readFileSync('package.json'))
+  }
+
+  const expected = cloneDeep(original)
+
+  if (expected.dependencies) {
+    Object.keys(expected.dependencies).forEach(function (k) {
+      if (k.indexOf('@bower_components') >= 0) {
+        delete expected.dependencies[k]
+      }
+    })
+  }
 
   const pkgs = fs.readdirSync(componentsDir)
 
@@ -152,10 +175,76 @@ async function main () {
     expected.dependencies[`@bower_components/${pkg}`] = `${source}#${target}`
   }
 
-  console.log(JSON.stringify(expected, null, 2))
+  if (typeof expected.engines !== 'object') {
+    expected.engines = {}
+  }
+
+  if (!expected.engines.yarn) {
+    expected.engines.yarn = '>= 1.0.0'
+  }
+
+  if (!deepIs(expected, original)) {
+    if (cli.flags.diff) {
+      if (process.stdout.isTTY) {
+        console.log(difflet.compare(original, expected, { indent: 2 }))
+      }
+
+      process.exit(0)
+    } else if (cli.flags.apply) {
+      fs.writeFileSync('package.json', JSON.stringify(expected, null, 2) + '\n')
+    } else {
+      if (fs.existsSync('package.json')) {
+        step('Update package.json', [
+          'Changes need to be made in package.json. Please run following to preview them:',
+          '',
+          '$ bower2yarn --diff',
+          '',
+          'And then apply them by running:',
+          '',
+          '$ bower2yarn --apply'
+        ])
+      }
+    }
+  }
+
+  if (!fs.existsSync(path.join(cwd, 'node_modules', '@bower_components'))) {
+    step('Install dependencies with Yarn', [
+      'Now what package.json contains Bower dependencies, please install them with:',
+      '$ yarn',
+      '',
+      'If you encounter issues during installation, please try:',
+      '$ yarn --ignore-engines --ignore-scripts',
+      '',
+      'You can use this command from now on to install both npm and bower dependencies.'
+    ])
+  }
+
+  const source = path.relative(process.cwd(), componentsDir)
+  const target = path.relative(
+    path.dirname(componentsDir),
+    'node_modules/@bower_components'
+  )
+  const script = 'rm -rf ' + source + ' && ln -sf ' + target + ' ' + source
+
+  if (fs.existsSync(path.join(cwd, 'bower.json'))) {
+    step('Remove bower.json and old bower components directory', [
+      'Your project is now converted to Yarn!',
+      '',
+      'You should find all bower components in node_modules/@bower_components',
+      '',
+      'If you would like them to be available in previous location',
+      'please consider adding following as postinstall script in package.json:',
+      '',
+      '"postinstall": "' + script + '"',
+      '',
+      'As a last step, please remove bower.json, .bowerrc',
+      'and point your scripts to node_modules/@bower_components instead'
+    ])
+  }
 
   step(
-    ['Your project is now converted to Yarn. Thank you for using Bower!'],
+    'Done',
+    ['Your project is now converted to Yarn! Thank you for using Bower!'],
     true
   )
 }
